requirements.txt
numpy
matplotlib
Pillow
scipy
scikit-image
gym
pybullet
